Update BSE.py
To keep memory-usage within limits on very long runs, added self.blotter_length to Trader class definition and self.tape_length to Orderbook definition: these set the maximum length of a trader's blotter records, and the exchange's/orderbook's tape records, respectively. Now when items are appended to blotter and/or tape, the lists are right-truncated to keep them within the length limit.