Update BSE.py
Bugfix: "s"->"s_index" at line 1572.

The bug was unlikely to result in any change of macro-level behavior: all strategies are initialized to 'evaluated':False, so when the buggy code was called with s!=s_index the worse that would happen is that a previously evaluated strategy would be incorrectly marked as evaluated=False; and so the test for all_evaluated at lines 1593-1597 could then fail when it should have succeeded if in fact all strategies had been evaluated -- but if that happens then the else clause commencing line 1647 would go through a fresh round of strategy evaluations. So the buggy code could potentially be inefficient, because it might unnecessarily repeat some strategy evaluations, slowing the rate of adaptation, but it does still adapt over time.