Added two data-files for specifying price offsets
Added two data-files offset-copper-1m-170912.csv and offset-ibm-1m-170831.csv to the repo as examples of the kind of freal-world financial data that can be used to specify dynamic changes in the BSE equilibrium price over the course of a market session. How this can be done is explained in the Wiki page "4. Altering the market supply and demand".